+++ scylla/src/transport/connection_pool.rs
@@ -1,7 +1,7 @@
 #[cfg(feature = "cloud")]
 use crate::cloud::set_ssl_config_for_scylla_cloud_host;
 
-use crate::errors::{ConnectionError, ConnectionPoolError, QueryError};
+use crate::errors::{BrokenConnectionErrorKind, ConnectionError, ConnectionPoolError, QueryError};
 use crate::routing::{Shard, ShardCount, Sharder};
 use crate::transport::{
     connection,
@@ -19,7 +19,6 @@ use super::NodeAddr;
 use arc_swap::ArcSwap;
 use futures::{future::RemoteHandle, stream::FuturesUnordered, Future, FutureExt, StreamExt};
 use rand::Rng;
-use scylla_cql::errors::BrokenConnectionErrorKind;
 use std::convert::TryInto;
 use std::num::NonZeroUsize;
 use std::pin::Pin;
